% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_role.R
\name{sagemaker_get_execution_role}
\alias{sagemaker_get_execution_role}
\title{Return the role ARN whose credentials are used to call the API.}
\usage{
sagemaker_get_execution_role()
}
\value{
(str): The role ARN
}
\description{
Return the role ARN whose credentials are used to call the API.
}
