% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sm_build}
\alias{sm_build}
\title{AWS CodeBuild and pushing to Amazon ECR}
\usage{
sm_build(
  repository = NULL,
  compute_type = c("BUILD_GENERAL1_SMALL", "BUILD_GENERAL1_MEDIUM",
    "BUILD_GENERAL1_LARGE", "BUILD_GENERAL1_2XLARGE"),
  role = NULL,
  dir = ".",
  bucket = NULL,
  vpc_id = NULL,
  subnet_ids = list(),
  security_group_ids = list(),
  log = TRUE,
  ...
)
}
\arguments{
\item{repository}{The ECR repository:tag for the image
(default: sagemaker-studio-${domain_id}:latest)}

\item{compute_type}{The CodeBuild compute type (default: BUILD_GENERAL1_SMALL)}

\item{role}{The IAM role name for CodeBuild to use (default: the Studio execution role).}

\item{dir}{Directory to build}

\item{bucket}{The S3 bucket to use for sending data to CodeBuild (if None,
use the SageMaker SDK default bucket).}

\item{vpc_id}{The Id of the VPC that will host the CodeBuild Project
(such as vpc-05c09f91d48831c8c).}

\item{subnet_ids}{List of subnet ids for the CodeBuild Project
(such as subnet-0b31f1863e9d31a67)}

\item{security_group_ids}{List of security group ids for
the CodeBuild Project (such as sg-0ce4ec0d0414d2ddc).}

\item{log}{Show the logs of the running CodeBuild build}

\item{...}{docker build parameters
\url{https://docs.docker.com/engine/reference/commandline/build/#options}
(NOTE: use "_" instead of "-" for example: docker optional parameter
\code{build-arg} becomes \code{build_arg})}
}
\description{
AWS CodeBuild and pushing to Amazon ECR
}
