% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{smdocker_config_log}
\alias{smdocker_config_log}
\title{smdocker logging system}
\usage{
smdocker_config_log(
  level = 3L,
  file = "",
  timestamp_fmt = "\%Y-\%m-\%d \%H:\%M:\%OS3"
)
}
\arguments{
\item{level}{integer to determine the level logging threshold.
\itemize{
\item{4L :} {DEBUG}
\item{3L :} {INFO}
\item{2L :} {WARNING}
\item{1L :} {ERROR}
}}

\item{file}{character path for logs to populate, default output logs to console.}

\item{timestamp_fmt}{character for timestamp format, see \code{\link[=format.POSIXct]{format.POSIXct()}}.}
}
\description{
Ability to configure smdocker logging system, through the use of smdocker
helper function \code{`smdocker_config_log`} or \code{R:base} \code{options} function.
Users are able to change logging levels without calling \code{smdocker.common} by
the use of \code{options} e.g. \code{options("smdocker.log_level" = 2L)}.
\itemize{
\item{smdocker.log_level} {integer: The minimum log level that should be tracked}
\item{smdocker.log_file} {
character: path for logs to populate, default output logs to console.
}
\item{smdocker.log_timestamp_fmt} {character: see \code{\link[=format.POSIXct]{format.POSIXct()}}}
}
}
\examples{
# log to a file
temp_file <- tempfile()
smdocker_config_log(file = temp_file)
unlink(temp_file)

# change log threshold to INFO
smdocker_config_log(level = 3L)

# reset to default config
smdocker_config_log()
}
